Changed the way ScheduleItems update.
When the Save button is clicked it triggers the save_edit method, which
first calls update_data.

update_data first makes sure that the new times the user entered are
valid.

If they are not, it returns false.
If they are, it updates all of the ScheduleItem attributes and then
calls update_display.

update_display updates the start and end time labels to match the times
entered in the date edits.
It also hides the end time if the duration is 0.

update_data then emits the updtaded signal with the new data.

Then it returns True.

If updat_data returned false, the save_data method then raises a
MessageBox telling the user the times entered are invalid. It returns.

If update_data is true, then save_data calls the stop_edits method which
returns the widget back to the default viewing state, replacing
DateEdits with Labels and making the desciption read-only.

I also modified DailyPlanner.py so it sorts the list of schedule items
in display_items instead of load_schedule.

Previously I only sorted the items in load_schedule, right after reading
all of the items from the database. This was an issue because items
would not be sorted to reflect any updates made until the load_schedule
method was called again.